Move test to step performers
